+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -132,7 +132,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-06T13:25:31.849Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-08T14:08:53.489Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")