+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGSJ2K3VJ6A7XSKYFDRE5A0H"
+        "01KGYSBD15C1JZNSBVCB3B8XJG"
     )
     r"""The unique identifier of the data source"""