+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGSJ2K3Q6PR9MYB33YGS9DAX"
+        "01KGYSBD0VYVX45JN7WA055NE4"
     )
     r"""The unique identifier of the data source"""