+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGSJ2K3SJ5Y8PEZ1DMMERVM3"
+        "01KGYSBD0XX26NPQ84K3EGNRCR"
     )
     r"""The unique identifier of the data source"""