+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2093,7 +2093,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KGSJ2JEBYFJV76WHH5XC4QK0"
+    id: Optional[str] = "01KGYSBCC8GG0F7JJ84DQ6Q7JR"
 
     description: Optional[str] = None