+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGSJ2K3ZEVY6VXP9N63KA8HN"
+        "01KGYSBD17427904QCZB7MD12F"
     )
     r"""The unique identifier of the data source"""