+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -259,7 +259,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KGSJ2JZBK1W4M831SBMJTH9T"
+    id: Optional[str] = "01KGYSBCWJXP6B93CZKPJ8KX17"
 
     description: Optional[str] = None
 
@@ -1214,7 +1214,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KGSJ2JYW2WYP42V8M48W2783"
+        "tool_01KGYSBCWADXR2MRZ0E959B58Y"
     )
 
     display_name: Optional[str] = None
@@ -1345,7 +1345,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KGSJ2JYWA43DD0S0H48RBY5K"
+    id: Optional[str] = "01KGYSBCW9PQXB52ZCD1VBJS5C"
 
     description: Optional[str] = None
 
@@ -1476,7 +1476,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KGSJ2JYTWENX5Y2FX4R64C2H"
+        "tool_01KGYSBCW8VRXBFP2SCFMJ4CEY"
     )
 
     display_name: Optional[str] = None
@@ -1777,7 +1777,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KGSJ2JYSRZAYHW3P3V7YXGB2"
+        "tool_01KGYSBCW7F2CD2Q86EX22Z6VH"
     )
 
     display_name: Optional[str] = None
@@ -1972,7 +1972,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KGSJ2JYQD2MRYCVYS5ME4ABA"
+        "tool_01KGYSBCVZN5GN53R4BA1DDQT9"
     )
 
     display_name: Optional[str] = None
@@ -2171,7 +2171,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KGSJ2JYPVD1PHB8JQ8SZ3VH1"
+        "tool_01KGYSBCVXZ8685C3PA38G6B4B"
     )
 
     display_name: Optional[str] = None