+++ packages/orq-rc/RELEASES.md
@@ -5369,3 +5369,13 @@ Based on:
 - [python v4.3.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.35 - packages/orq-rc
+
+## 2026-02-08 14:11:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.0 (2.811.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.37] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.37] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.37 - packages/orq-rc